requirements/requirements.txt
psycopg2-binary
flask
sqlalchemy
scrapy
pandas
selenium
nltk